requirements.txt
Django==4.2.13
# Updated from 3.15.1 — minor bugfix release
djangorestframework==3.15.2
django-filter==23.5
# Updated from 2.9.9 — includes connection stability fixes
psycopg2-binary==2.9.10
elasticsearch-dsl==8.13.1
elasticsearch==8.13.1
python-decouple==3.8
# Updated from 24.11.0 — seed_demo uses hardcoded Swedish data pools, not Faker API
Faker==37.1.0
gunicorn==22.0.0
whitenoise==6.6.0
django-cors-headers==4.3.1
celery==5.4.0
redis==5.0.4
